Added basic version of domain crawler.
Crawler is based off of the node-crawler library. It is seeded with a set
of initial domains, and then crawls the entire domain reachable from the
index page. For each page downloaded, the crawler parses the link tags,
finds all the links that point to other pages in the same domain, and then
adds them to the crawling queue if they are not already present. The
crawler can only reach pages on the domain that are reachable via some
link path from the index page.

The crawler itself is implemented in domain_crawler.js, and the test.js
file provides a simple test using the http://www.tamu.edu domain.